+++ barnhunt/templating.py
@@ -5,7 +5,7 @@ import os
 import random
 import re
 
-from six import python_2_unicode_compatible
+from six import python_2_unicode_compatible, text_type
 
 from .layerinfo import FlaggedLayerInfo, LayerFlags
 from .inkscape import svg
@@ -163,7 +163,7 @@ def random_rats(context, n=5, min=1, max=5, seed=None):
 def safepath(path_comp):
     """A jinja filter to replace shell-unfriendly characters with underscore.
     """
-    return re.sub(r"[\000-\040/\\\177\s]", '_', path_comp,
+    return re.sub(r"[\000-\040/\\\177\s]", '_', text_type(path_comp),
                   flags=re.UNICODE)