+++ CHANGES.rst
@@ -5,8 +5,10 @@ Changes
 Next Release
 ============
 
-The template for the output filename has been generalized to work sensibly
-in the case where overlays are nested more than two deep.
+- Python 3.7 is now supported.
+
+- The template for the output filename has been generalized to work
+  sensibly in the case where overlays are nested more than two deep.
 
 Templating
 ----------